Pass user to CSV import and sign in, refs #12083
Fixed issue where, when doing a CSV import from the web UI, via an asynchronous
job, information object audit logging wasn't noting the user that created the
job.

Added, to QubitCsvImport, the relaying of the current logged-in user to to the
import task as a CLI option (if information objects are being imported).

Added CLI option to arBaseTask that allows a username to be specified and signs
this user in before executing the task.

Made changes to classes relating to CSV import of information objects so they
would work with these changes.